tests/wpt/web-platform-tests/css/css-ui/parsing/resize-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 3: getComputedStyle().resize</title>
<link rel="help" href="https://drafts.csswg.org/css-ui-3/#resize">
<meta name="assert" content="resize computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("resize", "none");
test_computed_value("resize", "both");
test_computed_value("resize", "horizontal");
test_computed_value("resize", "vertical");
</script>
</body>
</html>
